backend/analyzer/requirements-test.txt
# Зависимости для тестирования SIEM Analyzer

# HTTP клиент для API тестирования
aiohttp>=3.8.0

# Асинхронное тестирование
pytest-asyncio>=0.21.0

# Тестирование
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Валидация данных
pydantic>=2.0.0

# Работа с YAML
PyYAML>=6.0

# Логирование
colorlog>=6.7.0

# Утилиты для тестирования
freezegun>=1.2.0
factory-boy>=3.2.0

# Асинхронные утилиты
asyncio-mqtt>=0.13.0

# Для интеграционных тестов
httpx>=0.24.0
websockets>=11.0.0

# Для нагрузочного тестирования
locust>=2.15.0

# Для мониторинга производительности
memory-profiler>=0.61.0
psutil>=5.9.0
